Implement database with service

+++ app/src/main/java/com/provectus/public_transport/model/TransportRoutes.java
@@ -37,7 +37,7 @@ public class TransportRoutes {
         this.id = id;
     }
 
-    public long getNumber() {
+    public int getNumber() {
         return number;
     }
